requirements.txt
django
scrapy
djangorestframework
scrapyd 
python-scrapyd-api
requests